Win32: Try to fix the nightly build errors. (Directory went away?)
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@19580 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ packaging/win32/dist-impl.sh
@@ -128,8 +128,11 @@ function dist_gnome() {
     rm -rf $TMP_UDIR/gtk2_prefs-*
 
     rm -rf $DIST_UDIR/etc/gconf/gconf.xml.defaults/{desktop,schemas}
-    cp -a $DIST_UDIR/lib/locale $DIST_UDIR/share
-    rm -rf $DIST_UDIR/lib/locale
+    if [ -d $DIST_UDIR/lib/locale ] ; then
+        # Huh, is this removed in newer gtk?
+        cp -a $DIST_UDIR/lib/locale $DIST_UDIR/share
+        rm -rf $DIST_UDIR/lib/locale
+    fi
 }
 
 function dist_isocodes() {
